+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.1.9] - 2017-22-12
+### Added
+
+- Added support for comments in require dependency list. Special thanks to [prantlf](https://github.com/prantlf)
+- Fixed problem navigating to an inline require statement's module.
+
+
+## [0.1.7] - 2017-21-12
+### Added
+
+- Added RequireJS config file support. Special thanks to [prantlf](https://github.com/prantlf)
+
 ## [0.1.6] - 2017-29-08
 ### Changed